complete ithink

+++ source/online/gui/sidebar/tabs/HostServerTab.hx
@@ -4,9 +4,12 @@ import haxe.io.Path;
 import sys.FileSystem;
 import haxe.io.Error;
 
+//TODO force kill server and node.exe
+//TODO stop freezing
+
 class HostServerTab extends TabSprite {
     public function new() {
-        super('Host Server', 'server');
+        super('Host Server (EXPERIMENTAL)', 'server');
 		tabWidth = 600;
     }
 
@@ -26,7 +29,8 @@ class HostServerTab extends TabSprite {
 		addChild(startAndStop);
 
 		updateServer = new TabButton('update', () -> {
-			FileUtils.removeFiles('_server/');
+			if (FileSystem.exists('_server/'))
+				FileUtils.removeFiles('_server/');
 			prepareServer();
 		});
 		updateServer.x = startAndStop.x;
